+++ sentry-android-okhttp/src/test/java/io/sentry/android/okhttp/SentryOkHttpInterceptorTest.kt
@@ -12,6 +12,7 @@ import io.sentry.SentryOptions
 import io.sentry.SentryTraceHeader
 import io.sentry.SentryTracer
 import io.sentry.SpanStatus
+import io.sentry.TraceStateHeader
 import io.sentry.TransactionContext
 import java.io.IOException
 import kotlin.test.Test
@@ -40,7 +41,10 @@ class SentryOkHttpInterceptorTest {
         val sentryTracer = SentryTracer(TransactionContext("name", "op"), hub)
 
         init {
-            whenever(hub.options).thenReturn(SentryOptions())
+            whenever(hub.options).thenReturn(SentryOptions().apply {
+                dsn = "https://key@sentry.io/proj"
+                isTraceSampling = true
+            })
         }
 
         fun getSut(
@@ -73,11 +77,12 @@ class SentryOkHttpInterceptorTest {
                     .toMediaType())).url(fixture.server.url("/hello")).build() }
 
     @Test
-    fun `when there is an active span, adds sentry trace header to the request`() {
+    fun `when there is an active span, adds sentry trace headers to the request`() {
         val sut = fixture.getSut()
         sut.newCall(getRequest()).execute()
         val recorderRequest = fixture.server.takeRequest()
         assertNotNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])
+        assertNotNull(recorderRequest.headers[TraceStateHeader.TRACE_STATE_HEADER])
     }
 
     @Test
@@ -86,6 +91,7 @@ class SentryOkHttpInterceptorTest {
         sut.newCall(getRequest()).execute()
         val recorderRequest = fixture.server.takeRequest()
         assertNull(recorderRequest.headers[SentryTraceHeader.SENTRY_TRACE_HEADER])
+        assertNull(recorderRequest.headers[TraceStateHeader.TRACE_STATE_HEADER])
     }
 
     @Test