+++ sentry/src/main/java/io/sentry/ISpan.java
@@ -39,6 +39,26 @@ public interface ISpan {
   @NotNull
   SentryTraceHeader toSentryTrace();
 
+  /**
+   * Returns the trace state information. @see <a
+   * href="https://develop.sentry.dev/sdk/trace-context/">Trace Context</a>.
+   *
+   * @return a trace state or {@code null} if {@link SentryOptions#isTraceSampling()} is disabled.
+   */
+  @Nullable
+  @ApiStatus.Experimental
+  TraceState traceState();
+
+  /**
+   * Returns the trace state that can be sent as a "tracestate" header.
+   *
+   * @return TraceStateHeader or {@code null} if {@link SentryOptions#isTraceSampling()} is
+   *     disabled.
+   */
+  @Nullable
+  @ApiStatus.Experimental
+  TraceStateHeader toTraceStateHeader();
+
   /** Sets span timestamp marking this span as finished. */
   void finish();