+++ sentry/src/main/java/io/sentry/SentryEnvelopeHeader.java
@@ -13,10 +13,20 @@ public final class SentryEnvelopeHeader {
 
   private final @Nullable SdkVersion sdkVersion;
 
+  private final @Nullable TraceState trace;
+
   public SentryEnvelopeHeader(
       final @Nullable SentryId eventId, final @Nullable SdkVersion sdkVersion) {
+    this(eventId, sdkVersion, null);
+  }
+
+  public SentryEnvelopeHeader(
+      final @Nullable SentryId eventId,
+      final @Nullable SdkVersion sdkVersion,
+      final @Nullable TraceState trace) {
     this.eventId = eventId;
     this.sdkVersion = sdkVersion;
+    this.trace = trace;
   }
 
   public SentryEnvelopeHeader(final @Nullable SentryId eventId) {
@@ -33,5 +43,9 @@ public final class SentryEnvelopeHeader {
 
   public @Nullable SdkVersion getSdkVersion() {
     return sdkVersion;
+  }
+
+  public @Nullable TraceState getTrace() {
+    return trace;
   }
 }