+++ sentry/src/test/java/io/sentry/SentryTracerTest.kt
@@ -1,6 +1,7 @@
 package io.sentry
 
 import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.check
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
@@ -9,6 +10,7 @@ import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import io.sentry.protocol.App
 import io.sentry.protocol.Request
+import io.sentry.protocol.User
 import java.util.Date
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -27,6 +29,8 @@ class SentryTracerTest {
 
         init {
             options.dsn = "https://key@sentry.io/proj"
+            options.environment = "environment"
+            options.release = "release@3.0.0"
             hub = spy(Hub(options))
             hub.bindClient(mock())
         }
@@ -114,7 +118,7 @@ class SentryTracerTest {
         tracer.finish()
         verify(fixture.hub).captureTransaction(check {
             assertEquals(it.transaction, tracer.name)
-        })
+        }, anyOrNull())
     }
 
     @Test
@@ -143,7 +147,7 @@ class SentryTracerTest {
         tracer.finish()
         verify(fixture.hub).captureTransaction(check {
             assertEquals(request, it.request)
-        })
+        }, anyOrNull())
     }
 
     @Test
@@ -158,7 +162,7 @@ class SentryTracerTest {
             assertEquals(app, it.contexts.app)
             assertEquals("value", it.contexts["custom"])
             assertEquals(tracer.spanContext, it.contexts.trace)
-        })
+        }, anyOrNull())
     }
 
     @Test
@@ -171,7 +175,7 @@ class SentryTracerTest {
         verify(fixture.hub).captureTransaction(check {
             assertNotEquals(spanContext, it.contexts.trace)
             assertEquals(tracer.spanContext, it.contexts.trace)
-        })
+        }, anyOrNull())
     }
 
     @Test
@@ -297,7 +301,7 @@ class SentryTracerTest {
         verify(fixture.hub).setSpanContext(ex, transaction.root, "name")
         verify(fixture.hub).captureTransaction(check {
             assertEquals(transaction.root.spanContext, it.contexts.trace)
-        })
+        }, anyOrNull())
 
         assertEquals(SpanStatus.OK, transaction.status)
         assertEquals(timestamp, transaction.timestamp)
@@ -338,7 +342,7 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         transaction.startChild("op")
         transaction.finish()
-        verify(fixture.hub, never()).captureTransaction(any())
+        verify(fixture.hub, never()).captureTransaction(any(), any())
     }
 
     @Test
@@ -347,7 +351,7 @@ class SentryTracerTest {
         val child = transaction.startChild("op")
         child.finish()
         transaction.finish()
-        verify(fixture.hub).captureTransaction(any())
+        verify(fixture.hub).captureTransaction(any(), anyOrNull())
     }
 
     @Test
@@ -368,7 +372,7 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         val child = transaction.startChild("op")
         child.finish()
-        verify(fixture.hub, never()).captureTransaction(any())
+        verify(fixture.hub, never()).captureTransaction(any(), any())
     }
 
     @Test
@@ -376,10 +380,61 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         val child = transaction.startChild("op")
         transaction.finish(SpanStatus.INVALID_ARGUMENT)
-        verify(fixture.hub, never()).captureTransaction(any())
+        verify(fixture.hub, never()).captureTransaction(any(), any())
         child.finish()
         verify(fixture.hub, times(1)).captureTransaction(check {
             assertEquals(SpanStatus.INVALID_ARGUMENT, it.status)
+        }, anyOrNull())
+    }
+
+    @Test
+    fun `returns trace state`() {
+        val transaction = fixture.getSut({
+            it.isTraceSampling = true
         })
+        fixture.hub.setUser(User().apply {
+            id = "user-id"
+            others = mapOf("segment" to "pro")
+        })
+        val trace = transaction.traceState()
+        assertNotNull(trace) {
+            assertEquals(transaction.spanContext.traceId, it.traceId)
+            assertEquals("key", it.publicKey)
+            assertEquals("environment", it.environment)
+            assertEquals("release@3.0.0", it.release)
+            assertEquals(transaction.name, it.transaction)
+            assertNotNull(it.user) {
+                assertEquals("user-id", it.id)
+                assertEquals("pro", it.segment)
+            }
+        }
+    }
+
+    @Test
+    fun `trace state does not change once computed`() {
+        val transaction = fixture.getSut({
+            it.isTraceSampling = true
+        })
+        val traceBeforeUserSet = transaction.traceState()
+        fixture.hub.setUser(User().apply {
+            id = "user-id"
+        })
+        val traceAfterUserSet = transaction.traceState()
+        assertNotNull(traceAfterUserSet) {
+            assertEquals(it, traceBeforeUserSet)
+            assertNull(it.user)
+        }
+    }
+
+    @Test
+    fun `returns trace state header`() {
+        val transaction = fixture.getSut({
+            it.isTraceSampling = true
+        })
+        val header = transaction.toTraceStateHeader()
+        assertNotNull(header) {
+            assertEquals("tracestate", it.name)
+            assertNotNull(it.value)
+        }
     }
 }